Roles y permisos

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,9 +19,9 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('num_cedula');
+            $table->bigInteger('num_cedula');
             $table->string('direccion');
-            $table->integer('numero_telefono');
+            $table->bigInteger('numero_telefono');
             $table->string('municipio');
             $table->string('foto_perfil');
             $table->rememberToken();
